Cover direct build graph library typechecking
